src: generalize the expand-pos-to-coordinates task
