+++ legate_sparse_cpp.cmake
@@ -107,6 +107,7 @@ list(APPEND legate_sparse_SOURCES
   src/sparse/array/conv/coo_to_dense.cc
   src/sparse/array/conv/csr_to_dense.cc
   src/sparse/array/conv/dense_to_csr.cc
+  src/sparse/array/conv/pos_to_coordinates.cc
   src/sparse/array/csc/spmv.cc
   src/sparse/array/csr/add.cc
   src/sparse/array/csr/get_diagonal.cc
@@ -124,6 +125,7 @@ if(Legion_USE_OpenMP)
   list(APPEND legate_sparse_SOURCES
     src/sparse/array/conv/csr_to_dense_omp.cc
     src/sparse/array/conv/dense_to_csr_omp.cc
+    src/sparse/array/conv/pos_to_coordinates_omp.cc
     src/sparse/array/csc/spmv_omp.cc
     src/sparse/array/csr/add_omp.cc
     src/sparse/array/csr/get_diagonal_omp.cc
@@ -140,6 +142,7 @@ if(Legion_USE_CUDA)
   list(APPEND legate_sparse_SOURCES
     src/sparse/array/conv/csr_to_dense.cu
     src/sparse/array/conv/dense_to_csr.cu
+    src/sparse/array/conv/pos_to_coordinates.cu
     src/sparse/array/csc/spmv.cu
     src/sparse/array/csr/get_diagonal.cu
     src/sparse/array/csr/add.cu
