+++ src/tasks.h
@@ -231,91 +231,6 @@ class SortedCoordsToCounts : public SparseTask<SortedCoordsToCounts> {
 #endif
 };
 
-class ExpandPosToCoordinates : public SparseTask<ExpandPosToCoordinates> {
- public:
-  static const int TASK_ID = LEGATE_SPARSE_EXPAND_POS_TO_COORDINATES;
-  static void cpu_variant(legate::TaskContext& ctx);
-#ifdef LEGATE_USE_OPENMP
-  static void omp_variant(legate::TaskContext& ctx);
-#endif
-#ifdef LEGATE_USE_CUDA
-  static void gpu_variant(legate::TaskContext& ctx);
-#endif
- public:
-  // We make this struct templated so that we can define it here and have it
-  // instantiated (with the cuda annotations) where it is needed. Unfortunately
-  // we have to make this struct public for NVCC to be happy.
-  template <typename T>
-  struct volume : public thrust::unary_function<T, size_t> {
-#if defined(__CUDACC__)
-    __host__ __device__
-#endif
-      size_t
-      operator()(Legion::Rect<1> x)
-    {
-      return x.volume();
-    }
-  };
-
- private:
-  // We'll use thrust for the body of this function (for all variants),
-  // so we'll define a helper to do the heavy lifting of the task.
-  template <typename Policy>
-  static void expand_pos_impl(const Policy& policy,
-                              legate::AccessorRO<Legion::Rect<1>, 1> pos,
-                              Legion::Domain pos_domain,
-                              legate::AccessorWO<coord_ty, 1> result,
-                              Legion::Domain result_domain,
-                              Legion::Memory::Kind tempMemKind)
-  {
-    // Return early if there isn't any work to do. Entering this code
-    // with an empty domain results in CUDA errors for the thrust backend.
-    if (pos_domain.empty() || result_domain.empty()) return;
-
-    // This implementation of expand was inspired from
-    // https://huggingface.co/spaces/ma-xu/LIVE/blob/main/thrust/examples/expand.cu.
-    Legion::DeferredBuffer<size_t, 1> volumes({0, pos_domain.get_volume() - 1}, tempMemKind);
-    Legion::DeferredBuffer<size_t, 1> offsets({0, pos_domain.get_volume() - 1}, tempMemKind);
-    // Initialize all of our arrays.
-    thrust::fill(policy, volumes.ptr(0), volumes.ptr(0) + pos_domain.get_volume(), size_t(0));
-    thrust::fill(policy, offsets.ptr(0), offsets.ptr(0) + pos_domain.get_volume(), size_t(0));
-    thrust::fill(policy,
-                 result.ptr(result_domain.lo()),
-                 result.ptr(result_domain.lo()) + result_domain.get_volume(),
-                 coord_ty(0));
-    // Transform each pos rectangle into its volume. We have to make a
-    // temporary here because not all of the thrust functions accept a
-    // transform.
-    thrust::transform(policy,
-                      pos.ptr(pos_domain.lo()),
-                      pos.ptr(pos_domain.lo()) + pos_domain.get_volume(),
-                      volumes.ptr(0),
-                      volume<Legion::Rect<1>>{});
-    // Perform an exclusive scan to find the offsets to write coordinates into.
-    thrust::exclusive_scan(
-      policy, volumes.ptr(0), volumes.ptr(0) + pos_domain.get_volume(), offsets.ptr(0));
-    // Scatter the non-zero counts into their output indices.
-    thrust::scatter_if(policy,
-                       thrust::counting_iterator<coord_ty>(0),
-                       thrust::counting_iterator<coord_ty>(pos_domain.get_volume()),
-                       offsets.ptr(0),
-                       volumes.ptr(0),
-                       result.ptr(result_domain.lo()));
-    // Compute a max-scan over the output indices, filling in holes.
-    thrust::inclusive_scan(policy,
-                           result.ptr(result_domain.lo()),
-                           result.ptr(result_domain.lo()) + result_domain.get_volume(),
-                           result.ptr(result_domain.lo()),
-                           thrust::maximum<coord_ty>{});
-    // Gather input values according to the computed indices.
-    thrust::gather(policy,
-                   result.ptr(result_domain.lo()),
-                   result.ptr(result_domain.lo()) + result_domain.get_volume(),
-                   thrust::counting_iterator<coord_ty>(pos_domain.lo()[0]),
-                   result.ptr(result_domain.lo()));
-  }
-};
-
 // Tasks for conversion to dense arrays.
 
 class CSCToDense : public SparseTask<CSCToDense> {
